Issue #620: Moved existing fuse components to the new fuse module
