+++ components/fuse/.gitignore
@@ -1,0 +1,4 @@
+/.project
+/.settings
+/.classpath
+/target
