+++ components/fuse/src/main/java/org/datacleaner/components/fuse/CoalesceUnit.java
@@ -17,7 +17,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.datacleaner.beans.coalesce;
+package org.datacleaner.components.fuse;
 
 import java.util.Arrays;
 import java.util.List;
