+++ components/fuse/src/main/java/org/datacleaner/components/fuse/CoalesceUnitConverter.java
@@ -17,7 +17,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.datacleaner.beans.coalesce;
+package org.datacleaner.components.fuse;
 
 import org.datacleaner.api.Converter;
 import org.datacleaner.util.convert.ArrayConverter;
