+++ components/fuse/src/main/resources/org/datacleaner/components/fuse/CoalesceMultipleFieldsTransformer.png
