+++ components/fuse/src/test/java/org/datacleaner/beans/coalesce/CoalesceUnitConverterTest.java
@@ -24,6 +24,8 @@ import java.util.Arrays;
 import junit.framework.TestCase;
 
 import org.datacleaner.api.InputColumn;
+import org.datacleaner.components.fuse.CoalesceUnit;
+import org.datacleaner.components.fuse.CoalesceUnitConverter;
 import org.datacleaner.configuration.DataCleanerConfigurationImpl;
 import org.datacleaner.configuration.InjectionManagerImpl;
 import org.datacleaner.data.MockInputColumn;
